test: refactor tests folder
